sync update

+++ programs/luna-os/kernel/audio.h
@@ -1,2 +1,2 @@
-extern void play_sound(void* buffer, int size, int block);
+extern void play_sound(void* buffer, long int size, short short int block);
 extern void* CRASH_SOUND;
