+++ programs/luna-os/kernel/lufs.h
@@ -1,2 +1,2 @@
-extern void lufs_create_file(char* name, int size);
+extern void lufs_create_file(char* name, long int size);
 extern void lufs_write_file(char* name, char* buffer);
